Move MDM agent's CBMA control to own module
Jira-ID: 3290
